YQL-19747 Extract methods to syntax/token.h
